+++ src/output.rs
@@ -2,6 +2,7 @@ use std::{path::Path, time::Duration};
 
 use similar::{Algorithm, ChangeTag, TextDiff};
 
+use crate::content::yaml;
 use crate::snapshot::{MetaData, Snapshot};
 use crate::utils::{format_rust_expression, style, term_width};
 
@@ -113,30 +114,34 @@ pub fn print_snapshot_summary_with_title(
     println!("{title:━^width$}", title = "", width = width);
 }
 
+fn print_line(width: usize) {
+    println!("{:─^1$}", "", width);
+}
+
 pub fn print_changeset(old: &str, new: &str, metadata: &MetaData, show_info: bool) {
     let width = term_width();
     let diff = TextDiff::configure()
         .algorithm(Algorithm::Patience)
         .timeout(Duration::from_millis(500))
         .diff_lines(old, new);
-    println!("{:─^1$}", "", width);
+    print_line(width);
 
     if show_info {
         if let Some(expr) = metadata.expression() {
             println!("Expression: {}", style(format_rust_expression(expr)));
-            println!("{:─^1$}", "", width);
+            print_line(width);
         }
 
         if let Some(descr) = metadata.description() {
             println!("{}", descr);
-            println!("{:─^1$}", "", width);
+            print_line(width);
         }
 
         if let Some(info) = metadata.private_info() {
-            let out = info.as_yaml();
+            let out = yaml::to_string(info);
             // TODO: does the yaml output always start with '---'?
             println!("{}", out.trim().strip_prefix("---").unwrap().trim_start());
-            println!("{:─^1$}", "", width);
+            print_line(width);
         }
     }