+++ src/snapshot.rs
@@ -6,7 +6,7 @@ use std::io::{BufRead, BufReader, Write};
 use std::path::{Path, PathBuf};
 use std::time::{SystemTime, UNIX_EPOCH};
 
-use crate::content::{self, Content};
+use crate::content::{self, json, yaml, Content};
 
 use once_cell::sync::Lazy;
 
@@ -37,13 +37,13 @@ impl PendingInlineSnapshot {
         }
     }
 
-    pub fn load_batch<P: AsRef<Path>>(p: P) -> Result<Vec<PendingInlineSnapshot>, Box<dyn Error>> {
+    pub fn load_batch(p: &Path) -> Result<Vec<PendingInlineSnapshot>, Box<dyn Error>> {
         let contents = fs::read_to_string(p)?;
 
         let mut rv: Vec<Self> = contents
             .lines()
             .map(|line| {
-                let value = Content::from_yaml(line)?;
+                let value = yaml::parse_str(line)?;
                 Self::from_content(value)
             })
             .collect::<Result<_, Box<dyn Error>>>()?;
@@ -56,10 +56,7 @@ impl PendingInlineSnapshot {
         Ok(rv)
     }
 
-    pub fn save_batch<P: AsRef<Path>>(
-        p: P,
-        batch: &[PendingInlineSnapshot],
-    ) -> Result<(), Box<dyn Error>> {
+    pub fn save_batch(p: &Path, batch: &[PendingInlineSnapshot]) -> Result<(), Box<dyn Error>> {
         fs::remove_file(&p).ok();
         for snap in batch {
             snap.save(&p)?;
@@ -67,9 +64,9 @@ impl PendingInlineSnapshot {
         Ok(())
     }
 
-    pub fn save<P: AsRef<Path>>(&self, p: P) -> Result<(), Box<dyn Error>> {
+    pub fn save(&self, p: &Path) -> Result<(), Box<dyn Error>> {
         let mut f = fs::OpenOptions::new().create(true).append(true).open(p)?;
-        let mut s = self.as_content().as_json();
+        let mut s = json::to_string(&self.as_content());
         s.push('\n');
         f.write_all(s.as_bytes())?;
         Ok(())
@@ -77,22 +74,24 @@ impl PendingInlineSnapshot {
 
     fn from_content(content: Content) -> Result<PendingInlineSnapshot, Box<dyn Error>> {
         if let Content::Map(map) = content {
-            let mut map = content::utils::into_unordered_struct_fields(map)?;
+            let mut run_id = None;
+            let mut line = None;
+            let mut old = None;
+            let mut new = None;
 
-            let run_id = content::utils::pop_str(&mut map, "run_id")?;
-            let line = content::utils::pop_u32(&mut map, "line")?;
-            let new = match map.remove("new") {
-                None | Some(Content::None) => None,
-                Some(non_null) => Some(Snapshot::from_content(non_null)?),
-            };
-            let old = match map.remove("old") {
-                None | Some(Content::None) => None,
-                Some(non_null) => Some(Snapshot::from_content(non_null)?),
-            };
+            for (key, value) in map.into_iter() {
+                match key.as_str() {
+                    Some("run_id") => run_id = value.as_str().map(|x| x.to_string()),
+                    Some("line") => line = value.as_u64().map(|x| x as u32),
+                    Some("old") if !value.is_nil() => old = Some(Snapshot::from_content(value)?),
+                    Some("new") if !value.is_nil() => new = Some(Snapshot::from_content(value)?),
+                    _ => {}
+                }
+            }
 
             Ok(PendingInlineSnapshot {
-                run_id,
-                line,
+                run_id: run_id.ok_or(content::Error::MissingField)?,
+                line: line.ok_or(content::Error::MissingField)?,
                 new,
                 old,
             })
@@ -187,14 +186,24 @@ impl MetaData {
 
     fn from_content(content: Content) -> Result<MetaData, Box<dyn Error>> {
         if let Content::Map(map) = content {
-            let mut map = content::utils::into_unordered_struct_fields(map)?;
+            let mut source = None;
+            let mut assertion_line = None;
+            let mut description = None;
+            let mut expression = None;
+            let mut info = None;
+            let mut input_file = None;
 
-            let source = content::utils::pop_nullable_str(&mut map, "source")?;
-            let assertion_line = content::utils::pop_nullable_u32(&mut map, "assertion_line")?;
-            let description = content::utils::pop_nullable_str(&mut map, "description")?;
-            let expression = content::utils::pop_nullable_str(&mut map, "expression")?;
-            let info = map.remove("info");
-            let input_file = content::utils::pop_nullable_str(&mut map, "input_file")?;
+            for (key, value) in map.into_iter() {
+                match key.as_str() {
+                    Some("source") => source = value.as_str().map(|x| x.to_string()),
+                    Some("assertion_line") => assertion_line = value.as_u64().map(|x| x as u32),
+                    Some("description") => description = value.as_str().map(Into::into),
+                    Some("expression") => expression = value.as_str().map(Into::into),
+                    Some("info") if !value.is_nil() => info = Some(value),
+                    Some("input_file") => input_file = value.as_str().map(Into::into),
+                    _ => {}
+                }
+            }
 
             Ok(MetaData {
                 source,
@@ -214,14 +223,14 @@ impl MetaData {
         if let Some(source) = self.source.as_deref() {
             fields.push(("source", Content::from(source)));
         }
-        if let Some(expression) = self.expression.as_deref() {
-            fields.push(("expression", Content::from(expression)));
-        }
         if let Some(line) = self.assertion_line {
             fields.push(("assertion_line", Content::from(line)));
         }
         if let Some(description) = self.description.as_deref() {
             fields.push(("description", Content::from(description)));
+        }
+        if let Some(expression) = self.expression.as_deref() {
+            fields.push(("expression", Content::from(expression)));
         }
         if let Some(info) = &self.info {
             fields.push(("info", info.to_owned()));
@@ -245,8 +254,8 @@ pub struct Snapshot {
 
 impl Snapshot {
     /// Loads a snapshot from a file.
-    pub fn from_file<P: AsRef<Path>>(p: P) -> Result<Snapshot, Box<dyn Error>> {
-        let mut f = BufReader::new(fs::File::open(p.as_ref())?);
+    pub fn from_file(p: &Path) -> Result<Snapshot, Box<dyn Error>> {
+        let mut f = BufReader::new(fs::File::open(p)?);
         let mut buf = String::new();
 
         f.read_line(&mut buf)?;
@@ -263,7 +272,7 @@ impl Snapshot {
                     break;
                 }
             }
-            let content = Content::from_yaml(&buf)?;
+            let content = yaml::parse_str(&buf)?;
             MetaData::from_content(content)?
         // legacy format
         } else {
@@ -300,7 +309,6 @@ impl Snapshot {
         }
 
         let module_name = p
-            .as_ref()
             .file_name()
             .unwrap()
             .to_str()
@@ -311,7 +319,6 @@ impl Snapshot {
             .to_string();
 
         let snapshot_name = p
-            .as_ref()
             .file_name()
             .unwrap()
             .to_str()
@@ -348,20 +355,33 @@ impl Snapshot {
 
     fn from_content(content: Content) -> Result<Snapshot, Box<dyn Error>> {
         if let Content::Map(map) = content {
-            let mut map = content::utils::into_unordered_struct_fields(map)?;
+            let mut module_name = None;
+            let mut snapshot_name = None;
+            let mut metadata = None;
+            let mut snapshot = None;
 
-            let module_name = content::utils::pop_str(&mut map, "module_name")?;
-            let snapshot_name = content::utils::pop_nullable_str(&mut map, "snapshot_name")?;
-            let metadata = MetaData::from_content(
-                map.remove("metadata").ok_or(content::Error::MissingField)?,
-            )?;
-            let snapshot = SnapshotContents(content::utils::pop_str(&mut map, "snapshot")?);
+            for (key, value) in map.into_iter() {
+                match key.as_str() {
+                    Some("module_name") => module_name = value.as_str().map(|x| x.to_string()),
+                    Some("snapshot_name") => snapshot_name = value.as_str().map(|x| x.to_string()),
+                    Some("metadata") => metadata = Some(MetaData::from_content(value)?),
+                    Some("snapshot") => {
+                        snapshot = Some(SnapshotContents(
+                            value
+                                .as_str()
+                                .ok_or(content::Error::UnexpectedDataType)?
+                                .to_string(),
+                        ))
+                    }
+                    _ => {}
+                }
+            }
 
             Ok(Snapshot {
-                module_name,
+                module_name: module_name.ok_or(content::Error::MissingField)?,
                 snapshot_name,
-                metadata,
-                snapshot,
+                metadata: metadata.ok_or(content::Error::MissingField)?,
+                snapshot: snapshot.ok_or(content::Error::MissingField)?,
             })
         } else {
             Err(content::Error::UnexpectedDataType.into())
@@ -404,17 +424,12 @@ impl Snapshot {
         &self.snapshot.0
     }
 
-    fn save_with_metadata<P: AsRef<Path>>(
-        &self,
-        path: P,
-        md: &MetaData,
-    ) -> Result<(), Box<dyn Error>> {
-        let path = path.as_ref();
+    fn save_with_metadata(&self, path: &Path, md: &MetaData) -> Result<(), Box<dyn Error>> {
         if let Some(folder) = path.parent() {
             fs::create_dir_all(&folder)?;
         }
         let mut f = fs::File::create(&path)?;
-        let blob = md.as_content().as_yaml();
+        let blob = yaml::to_string(&md.as_content());
         f.write_all(blob.as_bytes())?;
         f.write_all(b"---\n")?;
         f.write_all(self.contents_str().as_bytes())?;
@@ -424,7 +439,7 @@ impl Snapshot {
 
     /// Saves the snapshot.
     #[doc(hidden)]
-    pub fn save<P: AsRef<Path>>(&self, path: P) -> Result<(), Box<dyn Error>> {
+    pub fn save(&self, path: &Path) -> Result<(), Box<dyn Error>> {
         // we do not want to retain the assertion line on the metadata when storing
         // as a regular snapshot.
         if self.metadata.assertion_line.is_some() {
@@ -437,7 +452,7 @@ impl Snapshot {
     }
 
     /// Same as `save` but also holds information only relevant for `.new` files.
-    pub(crate) fn save_new<P: AsRef<Path>>(&self, path: P) -> Result<(), Box<dyn Error>> {
+    pub(crate) fn save_new(&self, path: &Path) -> Result<(), Box<dyn Error>> {
         self.save_with_metadata(path, &self.metadata)
     }
 }